services/simulator-service/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
httpx==0.27.2
prometheus-fastapi-instrumentator==7.0.0
kafka-python==2.1.2
pytest==8.3.3
